6-size.c

+++ 0x00-hello_world/6-size.c
@@ -15,6 +15,6 @@ printf("size of char: %lu byte(s)\n", (unsigned long)sizeof(a));
 printf("size of int: %lu byte(s)\n", (unsigned long)sizeof(b));
 printf("size of long int: %lu byte(s)\n", (unsigned long)sizeof(c));
 printf("size of long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
-printf("size of float: %lu byte(s)\n", (unsigned long)sizeof(f));
+printf("size of float: %f byte(s)\n", (float)sizeof(f));
 return (0);
 }
